ref(pkg): move kubernetes scraper
